fix: move testGitConnectionV1 to configuration.go
