+++ sqle/api/controller/v1/configuration_ce.go
@@ -7,11 +7,7 @@ import (
 	e "errors"
 	"github.com/actiontech/sqle/sqle/api/controller"
 	"github.com/actiontech/sqle/sqle/errors"
-	"github.com/actiontech/sqle/sqle/utils"
-	"github.com/go-git/go-git/v5/plumbing"
-	"github.com/go-git/go-git/v5/plumbing/storer"
 	"github.com/labstack/echo/v4"
-	"net/http"
 )
 
 var (
@@ -71,78 +67,4 @@ func testCodingAuditConfigV1(c echo.Context) error {
 
 func getScheduledTaskDefaultOptionV1(c echo.Context) error {
 	return controller.JSONBaseErrorReq(c, errCommunityEditionDoesNotSupportScheduledNotify)
-}
-
-func testGitConnectionV1(c echo.Context) error {
-	request := new(TestGitConnectionReqV1)
-	if err := controller.BindAndValidateReq(c, request); err != nil {
-		return controller.JSONBaseErrorReq(c, err)
-	}
-	repository, _, cleanup, err := utils.CloneGitRepository(c.Request().Context(), request.GitHttpUrl, request.GitUserName, request.GitUserPassword)
-	if err != nil {
-		return c.JSON(http.StatusOK, &TestGitConnectionResV1{
-			BaseRes: controller.NewBaseReq(nil),
-			Data: TestGitConnectionResDataV1{
-				IsConnectedSuccess: false,
-				ErrorMessage:       err.Error(),
-			},
-		})
-	}
-	defer func() {
-		cleanupError := cleanup()
-		if cleanupError != nil {
-			c.Logger().Errorf("cleanup git repository failed, err: %v", cleanupError)
-		}
-	}()
-	references, err := repository.References()
-	if err != nil {
-		return c.JSON(http.StatusOK, &TestGitConnectionResV1{
-			BaseRes: controller.NewBaseReq(nil),
-			Data: TestGitConnectionResDataV1{
-				IsConnectedSuccess: false,
-				ErrorMessage:       err.Error(),
-			},
-		})
-	}
-	branches, err := getBranches(references)
-	return c.JSON(http.StatusOK, &TestGitConnectionResV1{
-		BaseRes: controller.NewBaseReq(nil),
-		Data: TestGitConnectionResDataV1{
-			IsConnectedSuccess: true,
-			Branches:           branches,
-		},
-	})
-}
-
-func getBranches(references storer.ReferenceIter) ([]string, error) {
-	branches := make([]string, 0)
-	err := references.ForEach(func(ref *plumbing.Reference) error {
-		if ref.Type() == plumbing.HashReference {
-			branches = append(branches, ref.Name().Short())
-		}
-		return nil
-	})
-	if err != nil {
-		return branches, err
-	}
-	if len(branches) < 2 {
-		return branches, nil
-	}
-	// 第一个元素确认了默认分支名，需要把可以checkout的默认分支提到第一个元素
-	defaultBranch := "origin/" + branches[0]
-	defaultBranchIndex := -1
-	for i, branch := range branches {
-		if branch == defaultBranch {
-			defaultBranchIndex = i
-			break
-		}
-	}
-	if defaultBranchIndex == -1 {
-		return branches, nil
-	}
-	//1. 根据第一个元素，找到其余元素中的默认分支
-	//2. 如果找到，将找到的默认分支名移到第一个元素，并且删除原来的第一个元素。
-	branches[0], branches[defaultBranchIndex] = branches[defaultBranchIndex], branches[0]
-	branches = append(branches[:defaultBranchIndex], branches[defaultBranchIndex+1:]...)
-	return branches, nil
 }
